Fixes a bug where m_quad_edges was a C-style 6 slot array and we wrote up to the 12th position, overwriting the data of the boolean member variables and basically causing random behaviour all around. It was super fun to debug this, by the way

+++ qnode.cpp
@@ -145,7 +145,7 @@ void QNode::imageCallback(const sensor_msgs::ImageConstPtr &msg_left,
       left_imagedata->SetScalarTypeToUnsignedChar();
       left_imagedata->AllocateScalars();
 
-      std::vector<float> edges(12, -1);
+      std::vector<double> edges(12, -1);
 
       int leftmost_top = cols;
       int leftmost_top_id;
